backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Autenticación
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9

# Cliente HTTP para Airtable
httpx==0.26.0

# Utilidades
python-dotenv==1.0.1
